added support for a case in which a file was already exist in the target directory. Now it returns the data used to be there before running the installation

+++ WindowsInstaller/WindowsInstaller/WindowsInstallerTask/FileInstallerTask.cpp
@@ -3,19 +3,29 @@
 #include "Utils/File.h"
 #include "Utils/win32_utils.h"
 
-FileInstallerTask::FileInstallerTask(std::wstring source_file_path, std::wstring target_directory_path)
-	: source_file_path(source_file_path), target_directory_path(target_directory_path)
+FileInstallerTask::FileInstallerTask(const std::wstring& source_file_path, const std::wstring& target_directory_path)
+	: source_file_path(source_file_path), target_directory_path(target_directory_path), previous_data({false, {}})
 {}
 
 
 void FileInstallerTask::execute()
 {
-	//File file_to_install(source_file_path, GENERIC_ALL, CREATE_ALWAYS, FILE_ATTRIBUTE_NORMAL);
+	std::wstring file_name = source_file_path.substr(source_file_path.find_last_of(L"/\\") + 1);
 
-	//file_to_install.copy(target_directory_path);
+	std::wstring new_file_path = win32_utils::path_combine(target_directory_path, file_name); // TODO: maybe filesystem::path?
+	
+	try {
+		File target_file(new_file_path, GENERIC_ALL, OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL);
 
-	// TODO: make sure.. is that okay? 
-	win32_utils::copy_file(source_file_path, target_directory_path, true);
+		// File Exists:
+		previous_data.did_exist = true;
+		previous_data.data = target_file.read_entire_file();
+
+	} catch (...) {
+		// File does not exists, so nothing to do
+	}
+
+	win32_utils::copy_file(source_file_path, new_file_path, false);
 
 	// Log here
 
@@ -24,12 +34,25 @@ void FileInstallerTask::execute()
 void FileInstallerTask::rollback()
 {
 	try {
-		std::wstring file_name = source_file_path.substr(source_file_path.find_last_of(L"/\\") + 2);
+		std::wstring file_name = source_file_path.substr(source_file_path.find_last_of(L"/\\") + 1);
 		std::wstring new_file_path = win32_utils::path_combine(target_directory_path, file_name);
-		win32_utils::delete_file(new_file_path);
+
+		if (previous_data.did_exist) {
+			recover_previous_file(new_file_path);
+		}
+		else {
+			win32_utils::delete_file(new_file_path);
+		}
 	} catch (...) {
 		throw;
 	}
+}
+
+void FileInstallerTask::recover_previous_file(const std::wstring& new_file_path)
+{
+	File previous_file(new_file_path, GENERIC_ALL, CREATE_ALWAYS, FILE_ATTRIBUTE_NORMAL);
+
+	previous_file.write(previous_data.data);
 }
 
 // TODO: refactor
@@ -39,7 +62,5 @@ void from_json(const nlohmann::json& j, std::shared_ptr<FileInstallerTask>& task
 	std::string target_directory_path = j.at("target_directory_path").get<std::string>();
 	task = std::make_shared<FileInstallerTask>(std::wstring(source_file_path.begin(), source_file_path.end()),
 		std::wstring(target_directory_path.begin(), target_directory_path.end()));
-	//task->source_file_path = std::wstring(source_file_path.begin(), source_file_path.end());
-	//task->target_directory_path = std::wstring(target_directory_path.begin(), target_directory_path.end());
 }
